requirements.txt
altgraph==0.17
appdirs==1.4.3
argh==0.26.2
attrs==21.2.0
autopep8==1.5.4
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
commonmark==0.9.1
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
et-xmlfile==1.0.1
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.11.21
html5lib==1.0.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.8
iniconfig==1.1.1
ipaddr==2.2.0
jieba==0.42.1
lockfile==0.12.2
lxml==4.6.1
msgpack==0.6.2
neo4j==4.1.1
openpyxl==3.0.7
packaging==20.3
pep517==0.8.2
pluggy==0.13.1
progress==1.5
py==1.10.0
pycodestyle==2.6.0
pydantic==1.7.2
Pygments==2.8.1
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.1
pyparsing==2.4.6
pytest==6.2.4
python-docx==0.8.10
pytoml==0.1.21
pytz==2020.4
PyYAML==5.3.1
requests==2.22.0
retrying==1.3.3
rfc3986==1.4.0
rich==10.0.1
scapy==2.4.4
schedule==1.0.0
six==1.14.0
sniffio==1.2.0
toml==0.10.1
tqdm==4.51.0
typing-extensions==3.7.4.3
urllib3==1.25.8
webencodings==0.5.1
xlrd==2.0.1
xlwt==1.3.0
